socket_mode/requirements-socket.txt
slack-bolt
python-dotenv
tzdata